+++ tests/test_lin_solvers.py
@@ -18,6 +18,37 @@ class TestOptSolvers(unittest.TestCase):
 
         np.random.seed(2)
 
+    def test_umfpack(self):
+
+        A = np.random.randn(100,100)
+        b = np.random.randn(100)
+        
+        try:
+            umf = opt.lin_solver.new_linsolver('umfpack','unsymmetric')
+        except ImportError:
+            raise unittest.SkipTest('no umfpack')
+
+        self.assertTrue(isinstance(umf, opt.lin_solver.LinSolverUMFPACK))
+        umf.analyze(A)
+        umf.factorize(A)
+        x = umf.solve(b)
+
+        self.assertLess(norm(np.dot(A,x)-b),1e-10)
+
+    def test_superlu(self):
+
+        A = np.random.randn(100,100)
+        b = np.random.randn(100)
+
+        superlu = opt.lin_solver.new_linsolver('superlu','unsymmetric')
+        self.assertTrue(isinstance(superlu, opt.lin_solver.LinSolverSUPERLU))
+
+        superlu.analyze(A)
+        superlu.factorize(A)
+        x = superlu.solve(b)
+
+        self.assertLess(norm(np.dot(A,x)-b),1e-10)
+
     def test_mumps(self):
 
         A = np.random.randn(100,100)
@@ -28,6 +59,7 @@ class TestOptSolvers(unittest.TestCase):
         except ImportError:
             raise unittest.SkipTest('no mumps')
 
+        self.assertTrue(isinstance(mumps, opt.lin_solver.LinSolverMUMPS))
         mumps.analyze(A)
         mumps.factorize(A)
         x = mumps.solve(b)